fix bugs

+++ hive/src/main/java/com/netease/arctic/hive/catalog/ArcticHiveCatalog.java
@@ -130,18 +130,20 @@ public class ArcticHiveCatalog extends BasicArcticCatalog {
 
   @Override
   protected void doDropTable(TableMeta meta, boolean purge) {
-    // drop hive table operation will only delete hive table metadata
-    // delete data files operation will use BasicArcticCatalog
-    try {
-      hiveClientPool.run(client -> {
-        client.dropTable(meta.getTableIdentifier().getDatabase(),
-            meta.getTableIdentifier().getTableName(),
-            false /* deleteData */,
-            false /* ignoreUnknownTab */);
-        return null;
-      });
-    } catch (TException | InterruptedException e) {
-      throw new RuntimeException("Failed to drop table:" + meta.getTableIdentifier(), e);
+    if (HiveTableUtil.checkExist(hiveClientPool, TableIdentifier.of(meta.getTableIdentifier()))) {
+      // drop hive table operation will only delete hive table metadata
+      // delete data files operation will use BasicArcticCatalog
+      try {
+        hiveClientPool.run(client -> {
+          client.dropTable(meta.getTableIdentifier().getDatabase(),
+              meta.getTableIdentifier().getTableName(),
+              false /* deleteData */,
+              false /* ignoreUnknownTab */);
+          return null;
+        });
+      } catch (TException | InterruptedException e) {
+        throw new RuntimeException("Failed to drop table:" + meta.getTableIdentifier(), e);
+      }
     }
     super.doDropTable(meta, purge);
   }
